add event mock service and rename `compat-grpc` to `compat`
