+++ mock-collector/pom.xml
@@ -34,7 +34,7 @@
     <url>http://maven.apache.org</url>
 
     <properties>
-        <protocol.version>c3abc7e3d245ca8002dcc1f9bc93d45bc10bd382</protocol.version>
+        <protocol.version>7da226cfced7fa4eb91c6528e8c30827288531a0</protocol.version>
         <protocol.workingDirectory>${project.basedir}/target/protocol</protocol.workingDirectory>
         <protocol.repos>https://github.com/apache/skywalking-data-collect-protocol.git</protocol.repos>
         <protobuf-maven-plugin.version>0.6.1</protobuf-maven-plugin.version>
