+++ mock-collector/src/main/java/org/apache/skywalking/plugin/test/mockcollector/Main.java
@@ -20,15 +20,16 @@ package org.apache.skywalking.plugin.test.mockcollector;
 import io.grpc.netty.NettyServerBuilder;
 import io.netty.channel.local.LocalAddress;
 import org.apache.skywalking.plugin.test.mockcollector.mock.MockCLRMetricReportService;
+import org.apache.skywalking.plugin.test.mockcollector.mock.MockEventService;
 import org.apache.skywalking.plugin.test.mockcollector.mock.MockJVMMetricReportService;
 import org.apache.skywalking.plugin.test.mockcollector.mock.MockManagementService;
 import org.apache.skywalking.plugin.test.mockcollector.mock.MockMeterReportService;
 import org.apache.skywalking.plugin.test.mockcollector.mock.MockTraceSegmentService;
-import org.apache.skywalking.plugin.test.mockcollector.mock.compatgrpc.MockCLRMetricReportServiceCompat;
-import org.apache.skywalking.plugin.test.mockcollector.mock.compatgrpc.MockJVMMetricReportServiceCompat;
-import org.apache.skywalking.plugin.test.mockcollector.mock.compatgrpc.MockManagementServiceCompat;
-import org.apache.skywalking.plugin.test.mockcollector.mock.compatgrpc.MockMeterReportServiceCompat;
-import org.apache.skywalking.plugin.test.mockcollector.mock.compatgrpc.MockTraceSegmentServiceCompat;
+import org.apache.skywalking.plugin.test.mockcollector.mock.compat.MockCLRMetricReportServiceCompat;
+import org.apache.skywalking.plugin.test.mockcollector.mock.compat.MockJVMMetricReportServiceCompat;
+import org.apache.skywalking.plugin.test.mockcollector.mock.compat.MockManagementServiceCompat;
+import org.apache.skywalking.plugin.test.mockcollector.mock.compat.MockMeterReportServiceCompat;
+import org.apache.skywalking.plugin.test.mockcollector.mock.compat.MockTraceSegmentServiceCompat;
 import org.apache.skywalking.plugin.test.mockcollector.mock.rest.MockManagementServiceKeepAliveHandler;
 import org.apache.skywalking.plugin.test.mockcollector.mock.rest.MockManagementServiceReportPropertiesHandler;
 import org.apache.skywalking.plugin.test.mockcollector.mock.rest.MockTraceSegmentListCollectServletHandler;
@@ -54,6 +55,7 @@ public class Main {
                           .addService(new MockManagementService())
                           .addService(new MockTraceSegmentService())
                           .addService(new MockMeterReportService())
+                          .addService(new MockEventService())
                           .addService(new MockCLRMetricReportServiceCompat())
                           .addService(new MockJVMMetricReportServiceCompat())
                           .addService(new MockManagementServiceCompat())
